Test añadidos getByName y getByNameNotFound
